Adds a very rough first cut of a rare wizards plugin.
This code is brutal. I am not a javascripter and holy shit this is
confusing to me !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!

This commit adds a 'rare wizards' plugin. It watches Slack for messages
matching a regex for elftown images. If the image URL is new, it is
praised and added to the mongo db. If the URL has already been seen, the
mongo db object's occurrences count is incremented.

TODO:

1. Clean this the heck up.
2. The regex doesn't catch subfolders and misses some wizards
3. It would be cool to be able to spit back out a top 10 of the most
   common wizards

Thx @doeg for helping with my crappy JS.