Adds topic-watch plugin.
This commit adds a plugin that watches channels the bot is in for topic
changes. When a topic change occurs the new topic is saved along with
the ID of the user that changed it, the timestamp of the change and the
ID of the slack channel who's topic was changed.

*TODO*: Retrieving the history of topic changes/a random topic